modern-azure/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.20
pydantic==2.11.7
email-validator==2.2.0
python-dotenv==1.0.1
Werkzeug==3.0.4
requests==2.32.3
resend==2.5.1
